fix: home 페이지 오류 수정
